pyproject.toml
[project]
name = "snrnaseq-analysis"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "ehung", email = "ethanhung11@gmail.com" }
]
dependencies = [
    "pacmap==0.8.0",
    "scanpy>=1.11.2",
    "seaborn>=0.13.2",
    "glasbey>=0.3.0",
    "rpy2>=3.6.1",
    "scvi>=0.6.8",
    "doubletdetection>=4.3.0.post1",
    "anndata2ri>=1.3.2",
    "sc-toolbox>=0.12.3",
    "liana>=1.5.1",
    "pydeseq2>=0.5.1",
    "decoupler>=2.0.6",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=25.1.0",
    "notebook>=7.4.3",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/snrnaseq_analysis"]
